# book-network-ui/src/app/services/models/page-response-borrowed-books-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { BorrowedBooksResponse } from '../models/borrowed-books-response';
export interface PageResponseBorrowedBooksResponse {
  content?: Array<BorrowedBooksResponse>;
  first?: boolean;
  last?: boolean;
  number?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}